// src/app/interfaces/lot.interface.ts
export interface LotResp {
  lot: LotResult;
  status: string;
}

export interface LotResult {
  count: number;
  pages: number;
  page: number;
  limit: number;
  rows: Row[];
}

export interface Row {
  id?: number | undefined;
  quantity_liters: number;
  code: string;
  discount: number;
  discount_reason: string;
  cost_liter: number;
  reused_bottles: number;
  lot_total_cost: number;
  lot_type_id: number;
  company_id: number;
  status: boolean;
  createdAt: Date;
  updatedAt: Date;
  lot_type: LotType;
  spent_per_lots: spentPerLot[];
}

export interface LotType {
  id: number;
  name: string;
  status: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface spentPerLot {
  id: string;
  value: number;
  status: boolean;
  createAt: Date;
  updateAt: Date;
  spent: Expense;
}

export interface Expense {
  id: string;
  description: string;
  value: number;
  type: string;
  status: boolean;
  createAt: Date;
  updateAt: Date;
}

export interface Lot {
  id?: string;
  code: string;
  quantity_liters: number;
  discount: number;
  discount_reason: string;
  cost_liter: number;
  reused_bottles: number;
  lot_total_cost: number;
  lot_type_id: number;
  company_id: number;
}

export interface ReportLotExpenseProceeds {
  lotName: string[];
  expense: number[];
  proceeds: number[];
}

// Generated by https://quicktype.io

export interface ProceedsVsExpense {
  months: string[];
  proceedsValues: number[];
  expensesOperationalValues: number[];
  otherExpensesValues: number[];
  utilities: number[];
}